Wait for the device write to be completed before sending ack.
When sending messages from an application endpoint to a device
endpoint, delay the sending on the ack Messages (Error, RequestCompleted)
until the write is actually done on the device.

This avoids overflowing various parts of the system and
let us report actual errors to the client.

Next step is to put some upper bound on the size of the writing
queues.