Add NodeId: 128bits UUID such that nodes are uniquely identifiable
NodeId is a 128 bits UUID.
It is currently generated by the switch at device connection,
but the long term goal is to have that ID persisting on the device.
All Flatbuffers messages adressing a node use that type.

When talking to a thymio 2/wireless, only the first 2 bytes are
used, which should be enough entropy up to 100 robots.