Expose the WebSocket port on zeroconf.
While it make no sense to expose the websocket as a endpoint
as discovery do not work from web
(although it would work from node!), the launcher still need
to know the websocket endpoint to forward it to launch app.

this exposes the we-socket port associated to an endpoint as a
txt record (ws-port).