[Aseba] More lenient protocol version check
Dashel based-clients required the exact same protocol
version for both the device and the client.

This is not sensible because it prevents
updating the firmwares independently of the client.

Instead the protocol should be backward compatible
with older endpoint (and protocols 5,6 and 7 are).

This patch provides and check against
ASEBA_MAX_TARGET_PROTOCOL_VERSION, which is an arbitrary value
should a non backward-compatible protocol change ever be needed.