[TDM] Introduce scratchpads
Scratchpads are a mechanism to share source code between
applications.

Scratchpads are attached to a group ( so they are persistents)
and have an associated node id.

They are use to load AESL files.

Scratchpads changes can be monitor by applications.
They are not strictly tied to node:
The scratchpad code may from the code loaded on target.