Protect connection to non-localhost TDM with a password.
The TDM generates a random password (6 alpha numeric characters)
that is displayed in the interface of the localhost Launcher,
and broadcasted on the local network via zeroconf.

In effect this password is only yused by clients outside
of the local network.

The UI of the launcher is modified to add a password field.
Studio and VPL classic are modified to read the password
from the command line.

The JS api is modified to take a password a parameter,
in addition of the endpoint url.

createClient(url) becomes createClient(urlm, password)

Scratch, VPL3 and Blockly are NOT modified and need to be adapted.