Remote connection (#722)
* Transmit the server uuid and websocket port through the flatbufffer protocol

* Add a class to connect to a server using its address

* Dialog to connect to a remote server

* Pass the url to connect to to studio/vpl through the command line

* [TDM] Use a fixed port

* Protect connection to non-localhost TDM with a password.

The TDM generates a random password (6 alpha numeric characters)
that is displayed in the interface of the localhost Launcher,
and broadcasted on the local network via zeroconf.

In effect this password is only yused by clients outside
of the local network.

The UI of the launcher is modified to add a password field.
Studio and VPL classic are modified to read the password
from the command line.

The JS api is modified to take a password a parameter,
in addition of the endpoint url.

createClient(url) becomes createClient(urlm, password)

Scratch, VPL3 and Blockly are NOT modified and need to be adapted.

* move password display in remote connection view

* add checked box for automatic local network connection, status is store in settings

* Support remote connection on mac.

Because launching a bundle has no support for arguments,
we used an url instead.
This alternative merchanism to launch an application did not have
support for the new parameters (endpoint & password)

We modify the mac-specific code to support any argument,
then wire the new parameters in the Qt applications.

* Always connect to localhost, even if zero conf does not work.

The TDM always tries to initiate a connection to the local server,
so that if zero conf is not working ( server disabled, port blocked,
etc), everything can still works on the local machine.

* Use IP filtering to detect remote clients

Clients from the same subnet can connect to the TDM
without password.

The password is no longer broadcasted on zeroconf

* Let the TDM run even when the avahi deamom is missing

* Fix Windows build

* Extract password from handshake message

* Fix ip v4 mask computation on windows

* Hide the avahi error message when a local client is connected

* Wire the checkbox to let the user disable remote connections

* Disables both remote and network-local connections via a
TDM command line switch. Changing the option restarts the TDM

* Add a delay between the tdm process start and the
local connection attempt

* Minor typos fixes

* Add backup mechanism to display the IP

If the IP cannot be displayed, display
a link to an external website that shows the ip.

Also do not try to show a password before the local endpoint
has started.

* Fix detecting endpoint on same network

* The code was always checking for localhost
* IPV6 ips that could be mapped to IPv4 were not tested
against IPv4 interfaces

* Fix checkbox being inverted

* Handling TDM restart while it has not fully started yet.

* Fix empty password.

* Sometimes the TDM would generate an empty password because of incorrect
generation code.
* Sometimes the client would not receive the password due to the UI
not refreshing

* Cleanup logs

* Update JS API version

* Update blockly

* runtim 5.15 for flatpak

* Fix displaying password.

There were a few issues:

* The launcher was still trying to read the password out of the
zeroconf record. Because this is no longer broadcasted by the TDM,
the password was always set to empty once a discovery record was found.

* The signal that a local connection was established before the
endpoint was fully registered.

* upgrade scratch

* Revert "runtim 5.15 for flatpak"

This reverts commit 58716bcf233d5c1d7e8be9fb20aaf03b2fabf5ae.

* fix scratch and blockly

Co-authored-by: Michael Bonani <michael.bonani@mobsya.org>